fix: canvas panel renamed to side panel
