+++ packages/pentaho/src/Canvas/Canvas.mdx
@@ -7,7 +7,7 @@ import { Meta } from "@storybook/addon-docs";
 This section showcases all the Pentaho+ canvas components. At the moment, the following components are available:
 
 - [Bottom Panel](./?path=/docs/pentaho-canvas-bottom-panel--docs)
-- [Panel](./?path=/docs/pentaho-canvas-panel--docs)
+- [Side Panel](./?path=/docs/pentaho-canvas-side-panel--docs)
 - [Toolbar](./?path=/docs/pentaho-canvas-toolbar--docs)
 
 As a disclaimer, it's important to note that all canvas components are a work in progress and there might be breaking changes.
