--- packages/pentaho/src/Canvas/Panel/index.tsx
@@ -1,1 +1,0 @@
-export * from "./Panel";
