+++ packages/pentaho/src/Canvas/SidePanel/index.tsx
@@ -1,0 +1,1 @@
+export * from "./SidePanel";
