+++ packages/pentaho/src/Canvas/index.ts
@@ -1,5 +1,5 @@
 export * from "./Toolbar";
-export * from "./Panel";
+export * from "./SidePanel";
 export * from "./Tabs";
 export * from "./BottomPanel";
 export * from "./Tab";
